feat: add alert dialog based on app dialog
